evm(registerERC20NativeToken): reject long strings
This is related to #1509, but it probably does not completely fix it:
solo test passes but a similar cluster test will not; something wrong
happens when a contract produces a long error message in the receipt. We
should add the cluster test as soon as the new consensus is working.