fix(vm): prevent VM crash when depositing NFT with min storage deposit
This addresses #2067. If the chain has enough assets in the common
account, then the request is skipped instead of crashing.

TODO: If the chain does not have enough assets in the common account,
the VM still panics, probably because the required storage deposit for
the NFTOutput is not properly calculated.