feat(evm): accept non-view calls in eth_call
eth_call and eth_estimateGas now work in a similar fashion, executing a
VM task as if it was run by the consensus. As a consequence, non-view
calls can be executed, and also the block index/hash parameter is
respected.

Also, the BlockInfo object is refactored to store the current and previous
AliasOutput. In order to avoid resetting the testnet, the old schema is
supported; the new schema is in effect only for newly created blocks.

closes: #2105